+++ src/containerapp/HISTORY.rst
@@ -4,6 +4,7 @@ Release History
 ===============
 upcoming
 ++++++
+* 'az containerapp env update': Fix logs configuration about removing destination with `--logs-destination none`
 
 0.3.52
 ++++++